// redditclone-frontend/client-codegen-api/models/post-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Reddit Clone API
 * Clone Reddit With Spring Boot.
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface PostResponse
 */
export interface PostResponse {
    /**
     * 
     * @type {number}
     * @memberof PostResponse
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof PostResponse
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof PostResponse
     */
    url?: string;
    /**
     * 
     * @type {string}
     * @memberof PostResponse
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof PostResponse
     */
    userName?: string;
    /**
     * 
     * @type {string}
     * @memberof PostResponse
     */
    subredditName?: string;
    /**
     * 
     * @type {number}
     * @memberof PostResponse
     */
    voteCount?: number;
    /**
     * 
     * @type {number}
     * @memberof PostResponse
     */
    commentCount?: number;
    /**
     * 
     * @type {string}
     * @memberof PostResponse
     */
    duration?: string;
}