// redditclone-frontend/client-codegen-api/models/vote-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Reddit Clone API
 * Clone Reddit With Spring Boot.
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface VoteDto
 */
export interface VoteDto {
    /**
     * 
     * @type {string}
     * @memberof VoteDto
     */
    voteType?: VoteDtoVoteTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof VoteDto
     */
    postId?: number;
}

/**
    * @export
    * @enum {string}
    */
export enum VoteDtoVoteTypeEnum {
    UPVOTE = 'UPVOTE',
    DOWNVOTE = 'DOWNVOTE'
}